Refactor Rust intcode based solutions
Move each intcode based solution and related code into its own module.
